+++ modules/checkForUpdates.js
@@ -16,7 +16,7 @@ module.exports = (Callback) => {
         const parsedData = JSON.parse(rawData);
         const latestVersion = parsedData['dist-tags'].latest;
         if (global.version !== latestVersion) {
-          console.warn(`\n ${style.warningTitle('There is a newer version of MergerJS')} \n Please, run "merger update" to update.\n ${global.version} -> ${latestVersion }\n\n CHANGELOG: https://github.com/joao-neves95/merger-js/blob/master/CHANGELOG.md \n`);
+          console.warn(`\n ${style.warningTitle('There is a newer version of MergerJS')} \n Please, run "npm i merger-js -g" to update.\n ${global.version} -> ${latestVersion }\n\n CHANGELOG: https://github.com/joao-neves95/merger-js/blob/master/CHANGELOG.md \n`);
           notify('New version of MergerJS available.', `Please, run "merger update" to update.\n${global.version} -> ${latestVersion}`);
           return Callback();
         } else {