+++ modules/notifications.js
@@ -5,9 +5,6 @@ const progInfo = '\n\nMergerJS';
 // sound - Only Notification Center or Windows Toasters
 module.exports = {
   notif: (title, message) => {
-    if (global.config.notifications === undefined)
-      global.config.notifications = true;
-
     if (!global.config.notifications)
       return